refactor: clean up tsconfig, add path aliases, move type declarations
