+++ src/@types/BlacklistEntry.d.ts
@@ -1,4 +1,4 @@
-export interface BlacklistEntry {
+declare interface BlacklistEntry {
 	name: string;
 	uuid: string;
 	endDate: string;
