+++ src/events/mineflayer/chat/guildLevelUp.ts
@@ -1,5 +1,3 @@
-import { Event } from "../../../interfaces/Event";
-
 export default {
 	name: "chat:guildLevelUp",
 	runOnce: false,
