+++ src/events/mineflayer/chat/joinLimbo.ts
@@ -1,5 +1,3 @@
-import { Event } from "../../../interfaces/Event";
-
 export default {
 	name: "chat:joinLimbo",
 	runOnce: false,
