+++ src/events/mineflayer/chat/joinRequest.ts
@@ -1,8 +1,7 @@
-import { Event } from "../../../interfaces/Event";
-import fetchHypixelPlayerProfile from "../../../util/requests/fetchHypixelPlayerProfile";
-import fetchMojangProfile from "../../../util/requests/fetchMojangProfile";
-import isFetchError from "../../../util/requests/isFetchError";
-import isUserBlacklisted from "../../../util/blacklist/isUserBlacklisted";
+import fetchHypixelPlayerProfile from "@util/requests/fetchHypixelPlayerProfile";
+import fetchMojangProfile from "@util/requests/fetchMojangProfile";
+import isFetchError from "@util/requests/isFetchError";
+import isUserBlacklisted from "@util/blacklist/isUserBlacklisted";
 
 export default {
 	name: "chat:joinRequest",
