+++ src/events/mineflayer/chat/lobbyJoin.ts
@@ -1,5 +1,3 @@
-import { Event } from "../../../interfaces/Event";
-
 export default {
 	name: "chat:lobbyJoin",
 	runOnce: false,
