+++ src/events/mineflayer/chat/memberCount.ts
@@ -1,5 +1,3 @@
-import { Event } from "../../../interfaces/Event";
-
 export default {
 	name: "chat:memberCount",
 	runOnce: false,
