+++ src/events/mineflayer/chat/memberKick.ts
@@ -1,5 +1,4 @@
-import Emojis from "../../../util/emojis";
-import { Event } from "../../../interfaces/Event";
+import Emojis from "@util/emojis";
 import { escapeMarkdown } from "discord.js";
 
 export default {
