+++ src/events/mineflayer/chat/questTierComplete.ts
@@ -1,5 +1,4 @@
-import Emojis from "../../../util/emojis";
-import { Event } from "../../../interfaces/Event";
+import Emojis from "@util/emojis";
 
 export default {
 	name: "chat:questTierComplete",
