+++ src/events/mineflayer/handler/end.ts
@@ -1,5 +1,3 @@
-import { Event } from "../../../interfaces/Event";
-
 export default {
 	name: "end",
 	runOnce: false,
