+++ src/events/mineflayer/handler/error.ts
@@ -1,5 +1,3 @@
-import { Event } from "../../../interfaces/Event";
-
 export default {
 	name: "error",
 	runOnce: false,
