+++ src/index.ts
@@ -1,5 +1,5 @@
 import { config } from "dotenv";
 config();
 
-import Bot from "./classes/Bot";
+import Bot from "@classes/Bot";
 export default new Bot();
