--- src/interfaces/MojangProfileResponse.ts
@@ -1,4 +1,0 @@
-export interface MojangProfileResponse {
-	readonly id: string;
-	readonly name: string;
-}
