+++ src/util/blacklist/isUserBlacklisted.ts
@@ -1,4 +1,3 @@
-import { BlacklistEntry } from "../../interfaces/BlacklistEntry";
 import _blacklist from "./_blacklist.json";
 
 export default (uuid: string) => {
