+++ src/util/requests/FetchError.d.ts
@@ -1,4 +1,4 @@
-export interface FetchError {
+declare interface FetchError {
 	readonly status: number;
 	readonly statusText: string;
 }
