+++ src/util/requests/fetchMojangProfile.ts
@@ -1,8 +1,10 @@
-import { FetchError } from "../../interfaces/FetchError";
-import { MojangProfileResponse } from "../../interfaces/MojangProfileResponse";
-
 export default async (username: string) => {
 	const response = await fetch(`https://api.mojang.com/users/profiles/minecraft/${username}`);
 
 	return response.status === 200 ? ((await response.json()) as MojangProfileResponse) : (response as FetchError);
 };
+
+interface MojangProfileResponse {
+	readonly id: string;
+	readonly name: string;
+}
